tech(api): move POST /assessments/always-ok-validate-next-challenge endpoint to evaluation BC
