+++ api/src/shared/application/assessments/assessment-controller.js
@@ -1,15 +1,11 @@
 /**
  * @typedef {import('../../infrastructure/repositories/index.js').CertificationEvaluationRepository} CertificationEvaluationRepository
  */
-import * as certificationEvaluationApi from '../../../certification/evaluation/application/api/certification-evaluation-api.js';
 import { usecases as certificationUsecases } from '../../../certification/session-management/domain/usecases/index.js';
-import { Answer } from '../../../evaluation/domain/models/Answer.js';
 import { evaluationUsecases } from '../../../evaluation/domain/usecases/index.js';
 import * as competenceEvaluationSerializer from '../../../evaluation/infrastructure/serializers/jsonapi/competence-evaluation-serializer.js';
-import { usecases as questUsecases } from '../../../quest/domain/usecases/index.js';
 import { DomainTransaction } from '../../domain/DomainTransaction.js';
 import { sharedUsecases } from '../../domain/usecases/index.js';
-import * as assessmentRepository from '../../infrastructure/repositories/assessment-repository.js';
 import * as assessmentSerializer from '../../infrastructure/serializers/jsonapi/assessment-serializer.js';
 import { extractUserIdFromRequest, getChallengeLocale } from '../../infrastructure/utils/request-response-utils.js';
 
@@ -55,52 +51,6 @@ const findCompetenceEvaluations = async function (request) {
   return competenceEvaluationSerializer.serialize(competenceEvaluations);
 };
 
-const autoValidateNextChallenge = async function (request, h) {
-  const assessmentId = request.params.id;
-  const locale = getChallengeLocale(request);
-  const assessment = await assessmentRepository.getWithAnswers(assessmentId);
-  const userId = assessment.userId;
-  const fakeAnswer = new Answer({
-    assessmentId,
-    challengeId: assessment.lastChallengeId,
-    value: 'FAKE_ANSWER_WITH_AUTO_VALIDATE_NEXT_CHALLENGE',
-  });
-  if (assessment.isCompetenceEvaluation()) {
-    await evaluationUsecases.saveAndCorrectAnswerForCompetenceEvaluation({
-      answer: fakeAnswer,
-      assessment,
-      userId,
-      locale,
-      forceOKAnswer: true,
-    });
-  } else if (assessment.isForCampaign()) {
-    await evaluationUsecases.saveAndCorrectAnswerForCampaign({
-      answer: fakeAnswer,
-      assessment,
-      userId,
-      locale,
-      forceOKAnswer: true,
-    });
-  } else if (assessment.isCertification()) {
-    await certificationEvaluationApi.evaluateAndSaveAnswer({
-      answer: fakeAnswer,
-      userId,
-      certificationCourseId: assessment.certificationCourseId,
-      forceOKAnswer: true,
-    });
-  } else {
-    await evaluationUsecases.saveAndCorrectAnswerForDemoAndPreview({
-      answer: fakeAnswer,
-      assessment,
-      userId,
-      locale,
-      forceOKAnswer: true,
-    });
-  }
-  await questUsecases.rewardUser({ userId });
-  return h.response().code(204);
-};
-
 const createCertificationChallengeLiveAlert = async function (request, h) {
   const assessmentId = request.params.id;
   const challengeId = request.payload?.data?.attributes?.['challenge-id'];
@@ -112,7 +62,6 @@ const assessmentController = {
   getAssessmentWithNextChallenge,
   updateLastChallengeState,
   findCompetenceEvaluations,
-  autoValidateNextChallenge,
   createCertificationChallengeLiveAlert,
 };
