+++ api/src/shared/application/assessments/index.js
@@ -1,12 +1,8 @@
 import Joi from 'joi';
 
 import { assessmentAuthorization } from '../../../evaluation/application/pre-handlers/assessment-authorization.js';
-import { config } from '../../config.js';
 import { identifiersType } from '../../domain/types/identifiers-type.js';
-import { securityPreHandlers } from '../security-pre-handlers.js';
 import { assessmentController } from './assessment-controller.js';
-
-const { featureToggles } = config;
 
 const register = async function (server) {
   const routes = [
@@ -109,34 +105,6 @@ const register = async function (server) {
       },
     },
   ];
-
-  if (featureToggles.isAlwaysOkValidateNextChallengeEndpointEnabled) {
-    routes.push({
-      method: 'POST',
-      path: '/api/admin/assessments/{id}/always-ok-validate-next-challenge',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.assessmentId,
-          }),
-        },
-        handler: assessmentController.autoValidateNextChallenge,
-        tags: ['api'],
-      },
-    });
-  }
 
   server.route(routes);
 };
