+++ api/tests/evaluation/unit/application/assessments/assessment-routes_test.js
@@ -1,16 +1,14 @@
-import { expect } from 'chai';
 import sinon from 'sinon';
 
 import { assessmentController } from '../../../../../src/evaluation/application/assessments/assessment-controller.js';
 import * as moduleUnderTest from '../../../../../src/evaluation/application/assessments/assessment-routes.js';
+import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
+import { config as settings } from '../../../../../src/shared/config.js';
 import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
+import { expect } from '../../../../test-helper.js';
 import { HttpTestServer } from '../../../../tooling/server/http-test-server.js';
 
 describe('Evaluation | Unit | Application | assessment-routes', function () {
-  afterEach(function () {
-    sinon.restore();
-  });
-
   describe('POST /api/assessments', function () {
     it('should return 200', async function () {
       // given
@@ -47,6 +45,42 @@ describe('Evaluation | Unit | Application | assessment-routes', function () {
         expect(response.statusMessage).to.equal('Bad Request');
         sinon.assert.notCalled(assessmentController.save);
       });
+    });
+  });
+
+  describe('POST /api/admin/assessments/{id}/always-ok-validate-next-challenge', function () {
+    let originalEnvValue;
+
+    beforeEach(async function () {
+      originalEnvValue = settings.featureToggles.isAlwaysOkValidateNextChallengeEndpointEnabled;
+      settings.featureToggles.isAlwaysOkValidateNextChallengeEndpointEnabled = true;
+    });
+
+    afterEach(function () {
+      settings.featureToggles.isAlwaysOkValidateNextChallengeEndpointEnabled = originalEnvValue;
+    });
+
+    it('should return a response with an HTTP status code 403 if user does not have the rights', async function () {
+      // given
+      sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns((request, h) =>
+        h
+          .response({ errors: new Error('Unauthorized') })
+          .code(403)
+          .takeover(),
+      );
+
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+
+      // when
+      const { statusCode } = await httpTestServer.request(
+        'POST',
+        `/api/admin/assessments/123/always-ok-validate-next-challenge`,
+      );
+
+      // then
+      expect(securityPreHandlers.hasAtLeastOneAccessOf).to.have.be.called;
+      expect(statusCode).to.equal(403);
     });
   });
 });
